+++ src/azure/storage/blob/mod.rs
@@ -1,6 +1,12 @@
 mod list_blob_options;
 pub use self::list_blob_options::{ListBlobOptions, LIST_BLOB_OPTIONS_DEFAULT};
 
+mod put_block_options;
+pub use self::put_block_options::{PutBlockOptions, PUT_BLOCK_OPTIONS_DEFAULT};
+
+mod put_page_options;
+pub use self::put_page_options::{PutPageOptions, PUT_PAGE_OPTIONS_DEFAULT};
+
 use chrono::datetime::DateTime;
 use chrono::UTC;
 
@@ -9,7 +15,7 @@ use azure::storage::client::Client;
 
 use azure::core;
 use azure::core::{XMSRange, ContentMD5, XMSLeaseStatus, XMSLeaseDuration, XMSLeaseState,
-                  XMSLeaseId, XMSRangeGetContentMD5};
+                  XMSLeaseId, XMSRangeGetContentMD5, XMSClientRequestId};
 use azure::core::lease_id::LeaseId;
 use azure::core::parsing::{cast_must, cast_optional, from_azure_time, traverse};
 
@@ -36,6 +42,9 @@ use mime::Mime;
 use hyper::status::StatusCode;
 use hyper::header::{Headers, ContentType, ContentLength, LastModified, ContentEncoding,
                     ContentLanguage};
+
+use serialize::base64::{STANDARD, ToBase64, FromBase64};
+
 
 create_enum!(BlobType,
                             (BlockBlob,        "BlockBlob"),
@@ -490,21 +499,26 @@ impl Blob {
     pub fn put_page(&self,
                     c: &Client,
                     range: &BA512Range,
-                    lease_id: Option<LeaseId>,
+                    ppo: &PutPageOptions,
                     content: (&mut Read, u64))
                     -> Result<(), AzureError> {
 
-        let uri = format!("{}://{}.blob.core.windows.net/{}/{}?comp=page",
-                          c.auth_scheme(),
-                          c.account(),
-                          self.container_name,
-                          self.name);
+        let mut uri = format!("{}://{}.blob.core.windows.net/{}/{}?comp=page",
+                              c.auth_scheme(),
+                              c.account(),
+                              self.container_name,
+                              self.name);
+
+        if let Some(ref timeout) = ppo.timeout {
+            uri = format!("{}&timeout={}", uri, timeout);
+        }
+
         let mut headers = Headers::new();
 
         headers.set(XMSRange(range.into()));
         headers.set(XMSBlobContentLength(content.1));
-        if let Some(lease_id) = lease_id {
-            headers.set(XMSLeaseId(lease_id));
+        if let Some(ref lease_id) = ppo.lease_id {
+            headers.set(XMSLeaseId(lease_id.clone()));
         }
 
         headers.set(XMSPageWrite(PageWriteType::Update));
@@ -513,6 +527,45 @@ impl Blob {
                                               core::HTTPMethod::Put,
                                               &headers,
                                               Some(content)));
+        try!(core::errors::check_status(&mut resp, StatusCode::Created));
+
+        Ok(())
+    }
+
+    pub fn put_block(&self,
+                     c: &Client,
+                     block_id: &str,
+                     pbo: &PutBlockOptions,
+                     content: (&mut Read, u64))
+                     -> Result<(), AzureError> {
+
+        let mut uri = format!("{}://{}.blob.core.windows.net/{}/{}?comp=block&blockid={}",
+                              c.auth_scheme(),
+                              c.account(),
+                              self.container_name,
+                              self.name,
+                              block_id); //.to_base64(STANDARD));
+
+        if let Some(ref timeout) = pbo.timeout {
+            uri = format!("{}&timeout={}", uri, timeout);
+        }
+
+        let mut headers = Headers::new();
+
+        headers.set(XMSBlobContentLength(content.1));
+
+        if let Some(ref lease_id) = pbo.lease_id {
+            headers.set(XMSLeaseId(lease_id.clone()));
+        }
+        if let Some(ref request_id) = pbo.request_id {
+            headers.set(XMSClientRequestId(request_id.to_owned()));
+        }
+
+        let mut resp = try!(c.perform_request(&uri,
+                                              core::HTTPMethod::Put,
+                                              &headers,
+                                              Some(content)));
+
         try!(core::errors::check_status(&mut resp, StatusCode::Created));
 
         Ok(())